Create a user task based if the transition is mapped to a task.
After evaluating the transitions to see if it is mapped to a task
by calling the DMN. Kick off a process instance of the create task
workflow which will create a user task to be put on the user task
list. For now the task just has the CCD and task type on it.